Fix fluid export bus voiding fluids on crafting only mode
The Fluid Export Bus was voiding fluids on craft only mode because it
tried to craft it, and decided to continue with the normal export logic:
Extract fluid from ME system and fill the tank.

Unfortunately, there was a condition that prevented the extracted fluid
from going anywhere depending on craft only mode, so it got voided.

Commit fixes this by skipping the normal extract logic if set to craft
only. It also removes fluid simulation (we will see if they were actually
needed. Initial tests indicate no), which should improve performance by
a little bit.

Signed-off-by: Firenoo <49818773+firenoo@users.noreply.github.com>